% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{argraph}
\alias{argraph}
\title{The Ancestral Recombination Graph
This S3 class represents the ARG from the realized simulation}
\description{
A list of \code{bv_tree} for each discrete-loci, which constitutes the ARG
}
