% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{bvtreeToNewick}
\alias{bvtreeToNewick}
\title{Convert \code{bvtree} to Newick Tree Format}
\usage{
bvtreeToNewick(bvtree, tlim = 10)
}
\arguments{
\item{bvtree}{S3 class; internal class for the \code{polySimIBD}: package \link{bvtree}}

\item{tlim}{numeric; the maximum number of generations to consider before exiting gracefully if all samples have not coalesced}
}
\value{
Newick String
}
\description{
Recursively converts \code{bvtree} to Newick tree format for compatibility with other downstream packages
}
\details{
Note, the overall TMRCA is set to the \param{tlim}, which in Newick format looks inappropriately rooted
}
