% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{extract_haplotype_matrix}
\alias{extract_haplotype_matrix}
\title{Extract haplotypes from ARG}
\usage{
extract_haplotype_matrix(arg)
}
\arguments{
\item{arg}{set of bvtrees}
}
\value{
hapmat numeric matrix, such that each loci is represented by a row and an individual parasite (haplotype) is represented by a column
}
\description{
Extract haplotypes from ARG
}
