% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sWF.R
\name{get_arg}
\alias{get_arg}
\title{Get ancestral recombination graph from forward simulations}
\usage{
get_arg(swf, host_index = NULL, haplo_index = NULL)
}
\arguments{
\item{swf}{result of forwards simulation using the function \code{sim_swf()}}

\item{host_index}{a vector of target hosts. Defaults to all hosts}

\item{haplo_index}{a list of target haplotypes within the hosts specified by
\code{host_index}. Defaults to all haplotypes within the specified hosts}
}
\description{
Given an object \code{swf}, which is the result of forward
simulation using the function \code{sim_swf()}, walks backwards through the
ancestry and calculates the coalescent tree at every locus for the
specified haplotypes.
}
