% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_conn_from_root}
\alias{get_conn_from_root}
\title{Identify sub-trees recursively from root}
\usage{
get_conn_from_root(root, c)
}
\arguments{
\item{root}{internal identification of root indices}

\item{c}{connections from bvtree slot}
}
\description{
Internal function: Recursively loop through tree starting with root to find haplo-indices in the "bvtrees c slot" that are connected
}
\details{
Internal function, not intended for general use
}
