% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_conn_intervals}
\alias{get_conn_intervals}
\title{Get Connection Intervals}
\usage{
get_conn_intervals(uniqueconn, allconn)
}
\arguments{
\item{uniqueconn}{unique bvtree connections from the ARG}

\item{allconn}{all bvtree connections from the ARG}
}
\description{
Index where unique connections are in the entire genome for proper weighting
}
\details{
Internal function, not intended for general use
}
