% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_haplotype_matrix}
\alias{get_haplotype_matrix}
\title{Extract haplotypes from ARG}
\usage{
get_haplotype_matrix(arg)
}
\arguments{
\item{arg}{set of bvtrees}
}
\value{
hapmat numeric matrix; a matrix of multiallelic haplotypes for each parasite considered. Loci are in
rows and parasites (haplotypes) are in columns.
}
\description{
Extract haplotypes from ARG
}
