% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_pairwise_bv_ibd}
\alias{get_pairwise_bv_ibd}
\title{Calculate Between Host (pairwise) IBD by Interhost Relatedness}
\usage{
get_pairwise_bv_ibd(swf, host_index = NULL)
}
\arguments{
\item{swf}{result of forwards simulation using the function \code{sim_swf()}}

\item{host_index}{a vector of target hosts. Defaults to all hosts}
}
\value{
double of between-host IBD
}
\description{
Between host, or pairwise, IBD as presented in
Verity et al. 2020 (erity realized ibd (PMC7192906). The calculation
ignores intra-host relatedness and does not account for COI. Instead,
if there is any recent coalescence between any of the strains between
host 1 and host 2, the locus is condered to be IBD
}
