% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_pairwise_ibd}
\alias{get_pairwise_ibd}
\title{Effective IBD by Loci from ARG for a Pair of Hosts}
\usage{
get_pairwise_ibd(arg, host_index = NULL)
}
\arguments{
\item{host_index}{a vector of target hosts. Defaults to all hosts}
}
\value{

}
\description{
Given an object \code{ARG}, ***

Assumes that the minimum realized COI between the pairs of host determines
the denominator for the between realized IBD
}
\details{
Only accepts a pair of hosts. Ignores mutations as interrupting IBD segments.
}
