% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_pairwise_in_ibd}
\alias{get_pairwise_in_ibd}
\title{Calculate Between Host (pairwise) IBD by Interhost Relatedness}
\usage{
get_pairwise_in_ibd(swf, host_index = NULL)
}
\arguments{
\item{swf}{result of forwards simulation using the function \code{sim_swf()}}

\item{host_index}{a vector of target hosts. Defaults to all hosts}
}
\value{
double of between-host IBD
}
\description{
TODO
}
