% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_within_host_IBD}
\alias{get_within_host_IBD}
\title{Get Within-Host IBD from SWF Simulation}
\usage{
get_within_host_IBD(swf, host_index = NULL)
}
\arguments{
\item{swf}{result of forwards simulation using the function \code{sim_swf()}}

\item{host_index}{a vector of target hosts. Defaults to all hosts}
}
\value{
double of within-host IBD
}
\description{
TODO
}
\details{
Only accepts a single host
}
