% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_scripts.R
\name{join_simvcfs}
\alias{join_simvcfs}
\title{Join two vcfR objects that have the same CHROM/POS Columns}
\usage{
join_simvcfs(vcfRlist)
}
\arguments{
\item{vcflist}{list of vcfR object}
}
\description{
This assumes that we have two samples that came from the same original VCF and were split
for post-processing and are now being re-merged. Will not run if CHROMPOS for both samples are not identical
}
