% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation_model.R
\name{layer_mutations_on_ARG}
\alias{layer_mutations_on_ARG}
\title{Layer Mutations onto the ARG for Each Loci}
\usage{
layer_mutations_on_ARG(mutationrate, ARGsim)
}
\arguments{
\item{mutationrate}{numeric; the genome-wide per-generation mutation rate}

\item{ARGsim}{S4 object;}
}
\value{
GTmatrix numeric matrix;
}
\description{
Layer Mutations onto the ARG for Each Loci
}
