% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_scripts.R
\name{make_mixture}
\alias{make_mixture}
\title{make mixture}
\usage{
make_mixture(strains = list(), strains.prop = list(), CHROMPOS, dp,
  dpstd, samplename)
}
\arguments{
\item{strains}{named list; sample names with a vector of 0,2 corresponding to haplotypes}

\item{strains.prop}{numeric vector; proportion of the strains}

\item{CHROMPOS}{matrix <character, numeric>; genomic coordinates corresponding to chromosome and position}

\item{dp}{numeric; average read depth per loci}

\item{dpstd}{numeric; standard deviation around read depth}

\item{samplename}{character; sample name for resulting mixture}
}
\description{
Make mixture from list of samples
}
