% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_barcodes}
\alias{plot_barcodes}
\title{plot barcodes}
\usage{
plot_barcodes(hapmat, coi)
}
\arguments{
\item{hapmat}{matrix; a matrix of haplotypes}

\item{coi}{numeric vector; a vector of COIs per sample}
}
\description{
Produces a ggplot of barcode counts, in which width
indicates the number of times each barcode was seen, and all completely
unique barcodes (only seen once) are indicated in grey
}
