% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_tools.R
\name{plot_between_lineages_simulatedpedigree}
\alias{plot_between_lineages_simulatedpedigree}
\title{Plot Between Lineage IBDness from Simulated Pedigree}
\usage{
plot_between_lineages_simulatedpedigree(simulatedpedigree)
}
\description{
Plot Between Lineage IBDness from Simulated Pedigree
}
