% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_coalescence_trees}
\alias{plot_coalescence_trees}
\title{plot the bvtrees}
\usage{
plot_coalescence_trees(ARG, loci = NULL)
}
\arguments{
\item{ARG}{S4 object; The ARG object that was created from a structured Wright
Fisher Simulation via \code{sim_structured_WF} and then processed with \code{get_ARG}.}

\item{loci}{numeric vector; Loci to plot}
}
\value{
ggplot object of geom_segments
}
\description{
Produces a ggplot of bvtrees at every locus.
}
