% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_within_lineage_simulatedpedigree}
\alias{plot_within_lineage_simulatedpedigree}
\title{Plot Within Lineage IBD from Simulated Pedigree}
\usage{
plot_within_lineage_simulatedpedigree(sim, lineage)
}
\description{
Plot Within Lineage IBD from Simulated Pedigree
}
