% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plotARG}
\alias{plotARG}
\title{plot the ARG by loci from swf simulator}
\usage{
plotARG(
  swf,
  loci,
  host_index = 1:length(swf$coi),
  haplo_index = NULL,
  tlim = 999
)
}
\arguments{
\item{swf}{result of forwards simulation using the function \code{sim_swf()}}

\item{loci}{numeric vector; specific loci for plotting}

\item{host_index}{a vector of target hosts. Defaults to all hosts}

\item{haplo_index}{a list of target haplotypes within the hosts specified by
\code{host_index}. Defaults to all haplotypes within the specified hosts}

\item{tlim}{numeric; the maximum number of generations to consider before exiting gracefully if all samples have not coalesced}
}
\value{
a list of ggplots (layer: geom_segments) of the requested loci across selected hosts or haplotypes from the simulation
}
\description{
plot the ARG by loci from swf simulator
}
