% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{polySimIBD}
\alias{polySimIBD}
\title{Structured Wright Fisher Simulator for Malaria Genetics}
\description{
Unique features of the malaria life-cycle and transmission
dynamics requires extensions of typical population genetic simulators.
Using a spatial discrete-loci, discrete-time structured Wright Fisher model, we
simulate malaria population genetics forwards in time. Users are then able
to capture the full Ancestral Recombination Graph.
}
