% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sWF_hierarchical.R
\name{sim_hierarchical_structured_WF}
\alias{sim_hierarchical_structured_WF}
\title{The Structured Wright Fisher Model for IBD}
\usage{
sim_hierarchical_structured_WF(pos, N, MM, m, rho, mean_coi, tlim)
}
\arguments{
\item{pos}{dataframe <factor><numeric>; the genomic coordinates for chromosome and position of the sites}

\item{N}{numeric; The number of individuals to consider}

\item{m}{numeric; Probability of migration where m represents the probability of a parasite moving from one individual to another individual}

\item{rho}{numeric; expected recombination rate}

\item{mean_coi}{numeric; The lambda of a zero-truncated Poisson process, representing the average COI among inviduals}

\item{tlim}{numeric; the maximum number of generations to consider before exitting gracefully if all samples have not coalesced}
}
\value{
Returns a list of length G (the number of generations it took for all lineages to coalesce). Each
list item is a matrix with n x L dimensions where n is the number of parasites considered and L is loci.
Parasites are then indexed into demes (individuals) based on the COI distribution.
}
\description{
Simulate a population forwards with recombination that approximates the Structured Wright Fisher Process
and tracks haplotype identity by descent where individuals represent demes, such that within
a deme individual-level COI is considered
}
\details{
Data is simulated under the following framework:
\enumerate{
\item ... TODO>..
}
}
\examples{
\dontrun{

}

}
