% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sWF.R
\name{sim_structured_WF}
\alias{sim_structured_WF}
\title{The Structured Wright Fisher Model for IBD}
\usage{
sim_structured_WF(pos, K, m, rho, mean_coi)
}
\arguments{
\item{pos}{dataframe <factor><numeric>; the genomic coordinates for chromosome and position of the sites}

\item{K}{numeric; The number of individuals (e.g. demes) to consider}

\item{m}{numeric; Probability of migration where m represents the probability of moving from \eqn{deme_{origin}} to  \eqn{deme_{new}} by \eqn{m*(1-1/K)}}

\item{rho}{numeric; expected recombination rate}

\item{mean_coi}{numeric; The lambda of a right-shifted Poisson process, \eqn{1 + Pos(lambda)} representing the average COI of an individual deme}
}
\description{
Simulate a population forwards with recombination that approximates the Structured Wright Fisher Process
and tracks haplotype identity by descent where individuals represent demes, such that within
a deme individual-level COI is considered
}
\details{
Data is simulated under the following framework:
\enumerate{
\item Haploid genotypes are drawn from the \code{Nes}
\item Haploid genotypes are repeated based on the census size, \code{Nc}
\item Encode haplotypes with a \emph{int}
\item ... TODO>..
}

The forward in time simulation is performed by randomly drawing two strains from the previous generation,
where strains can either be homologous or heterologous haplotypes depending on the haplotype frequencies,
and allowing for mating to occur. Within the formed zygote, recombination events occur at a rate proportional
to the recombination rate and distance along the genome: \eqn{1-e^{-rho*d}}. This process is modeled with a
transition matrix and determines whether two samples enter a state of IBD or not. Once the recombination block
has been determined, the four potential haploid meiotic products are created.
TODO We then randomly pick one of the four which may not be ideal
}
