% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePLAF.R, R/vcf_scripts.R
\name{simPopData}
\alias{simPopData}
\alias{simPopData2vcfR}
\title{Simulate data ...todo...}
\usage{
simPopData(pos = c(sort(sample(1e+05, 100))), nsamps = 2, p = NULL,
  p_shape1 = 0.1, p_shape2 = 0.1, propMissing = 0)

simPopData2vcfR(simPopData, dp, dpstd)
}
\arguments{
\item{pos}{numeric vector; genomic coordinates}

\item{nsamps}{integer; number of samples}

\item{p_shape1}{integer; alpha in beta dist}

\item{p_shape2}{integer; beta in beta dist}

\item{propMissing}{integer; prob of missingness per sample}

\item{dp}{numeric; average read depth per loci}

\item{dpstd}{numeric; standard deviation around read depth}
}
\description{
Simulate a population of samples from an pop AF distribution

This converts an object of {simPopData} to a {vcfR} object
}
