% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePedigree.R
\name{simulate_IBD_pop_Pedigree}
\alias{simulate_IBD_pop_Pedigree}
\title{Simulate Pedigree IBD}
\usage{
simulate_IBD_pop_Pedigree(chrompos, PLAF, rho, k)
}
\arguments{
\item{chrompos}{matrix <factor><numeric>; the genomic coordinates for chromosome and position of the sites}
}
\description{
Simulate Pedigree IBD
}
