% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateWF.R
\name{simulate_IBD_pop_WF}
\alias{simulate_IBD_pop_WF}
\title{Simulate IBD-WF}
\usage{
simulate_IBD_pop_WF(chrompos, PLAF, Nes, rho, k)
}
\arguments{
\item{chrompos}{matrix <factor><numeric>; the genomic coordinates for chromosome and position of the sites}

\item{PLAF}{numeric vector; the population-level allele frequencies to simulate genotypes from}

\item{Nes}{numeric; The starting effective population size}

\item{rho}{numeric; expected recombination rate}

\item{k}{numeric; number of generation to simulate forward}
}
\description{
Simulate IBD-WF
}
\details{
Data is simulated under the following framework:
\enumerate{
\item Haploid genotypes are drawn from the \code{PLAF}
\item Encode haplotypes with a \emph{bit}
\item Simulate forward in time from the inital effective population \code{Nes} by drawing two samples
and allowing sexual recombination to occur.
}

The forward in time simulation is performed by randomly drawing two strains from the previous generation,
where strains can either be homologous or heterologous haplotypes depending on the haplotype frequencies,
and allowing for mating to occur. Within the formed zygote, recombination events occur at a rate proportional
to the recombination rate and distance along the genome: \eqn{1-e^{-rho*d}}. This process is modeled with a
transition matrix and determines whether two samples enter a state of IBD or not. Once the recombination block
has been determined, the four potential haploid meiotic products are created.
TODO We then randomly pick one of the four which may not be ideal
}
