% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sWF.R
\name{subset_bvtree}
\alias{subset_bvtree}
\title{Subset an object of class bvtree}
\usage{
subset_bvtree(bvtree, s)
}
\arguments{
\item{bvtree}{an object of class "bvtree"}

\item{s}{a vector specifying which elements in the bvtree to focus on}
}
\description{
Given a bvtree and a vector of indices \code{s}, creates a new
tree which is a subset of the original tree focusing only on the elements
\code{s}.
}
