% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{swfsim}
\alias{swfsim}
\title{The Structured Wright Fisher Model Output}
\description{
The realization of the Discrete-Time Discrete-Loci Spatial Wright Fisher Malaria Model contains all of the information
to create the ARG: each generation's parents, the resulting recombination events between parents, and the offspring haplotypes
}
\details{
This S3 class represents realization of the Discrete-Time Discrete-Loci Spatial Wright Fisher Malaria Model.
}
\section{Fields}{

\describe{
\item{\code{pos}}{vector; the genomic coordinates for chromosome and position of the sites}

\item{\code{coi}}{vector; the COI of each host}

\item{\code{recomb}}{list; recombination blocks for each ancestral host for each generation}

\item{\code{parent_host1}}{list; parent for host 1 for each generation}

\item{\code{parent_host2}}{list; parent for host 2 for each generation}

\item{\code{parent_haplo1}}{list; haplotypes for parent 1 for each generation}

\item{\code{parent_haplo2}}{list; haplotypes for parent 2 for each generation}
}}

