% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_trunc_poisson.R
\name{zero_trunc_poisson}
\alias{zero_trunc_poisson}
\title{Zero Truncated Poisson Model}
\usage{
zero_trunc_poisson(n, lambda)
}
\arguments{
\item{n}{numeric; number of draws to consider}

\item{lambda}{numeric; the lambda parameter in a poisson distribtuion}
}
\value{
integer vector of sucesses
}
\description{
Zero Truncated Poisson Model
}
