% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{argraph}
\alias{argraph}
\title{The Ancestral Recombination Graph}
\description{
A list of \code{bvtree} for each discrete-loci, which constitutes the ARG
}
\details{
This S3 class represents the ARG from the realized simulation
}
