% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{bvtreeToNewick}
\alias{bvtreeToNewick}
\title{Convert \code{bvtree} to Newick Tree Format}
\usage{
bvtreeToNewick(bvtree, tlim = 10)
}
\arguments{
\item{bvtree}{S3 class; internal class for the \code{polySimIBD}: package \code{bvtree}}

\item{tlim}{numeric; the maximum number of generations to consider before
exiting gracefully if all samples have not coalesced}
}
\value{
Newick String
}
\description{
Recursively converts \code{bvtree} to Newick tree format for compatibility with other downstream packages. The output format named leafs with distances that correspond to coalescent times. Additionally, clades have the total tree length explicitly (versus the acceptable Newick shorthand without total length).
}
\details{
Note, the overall time to the most recent common ancestor (TMRCA) is set to the \code{tlim}, which in Newick format looks inappropriately rooted.

Note, the function does not "know" the original \code{tlim} that was specified in the forward-simulation, and thus must be re-stated by the user.
}
