% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sWF_tools.R
\name{plot_coalescence_trees}
\alias{plot_coalescence_trees}
\title{plot the bvtrees}
\usage{
plot_coalescence_trees(ARGsim, loci)
}
\arguments{
\item{ARGsim}{S4 object;}

\item{loci}{numeric vector; loci which we want to plot}
}
\value{
ggplot object of geom_segments
}
\description{
plot the bvtrees
}
