% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatehmmIBD.R
\name{sim_hmmIBD}
\alias{sim_hmmIBD}
\title{Simulate IBD sections between two samples under the hmmIBD model}
\usage{
sim_hmmIBD(chrompos, PLAF, f, k, rho, pos)
}
\arguments{
\item{chrompos}{matrix <factor><numeric>; the genomic coordinates for chromosome and position of the sites}

\item{PLAF}{numeric vector; the population-level allele frequencies to simulate genotypes from}

\item{f}{the average relatedness between the two samples}

\item{k}{the number of generations separating the two lineages}

\item{rho}{the recombination rate}

\item{pos}{the genomic positions of the sites of interest

#'@details Data is simulated under the following framework:
\enumerate{
\item Draw two genotypes from the \code{PLAF}
\item Encode haplotypes with a \emph{bit}
\item Cross samples under the hmmIBD framework
}}
}
\description{
Walks along a vector of genomic locations and
swiches between two states that represent IBD and non-IBD using a Markov model.
The parameters that dictate the chance of switching state at any point include
the average level of relatedness (\code{f}), the physical distance between
loci in units of base pairs (from \code{pos}), and the recombination rate (\code{rho}),
which is assumed constant over all loci.
}
