% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forward_sWF.R
\name{sim_structured_WF}
\alias{sim_structured_WF}
\title{The Structured Wright Fisher Model for IBD}
\usage{
sim_structured_WF(pos, K, m, rho, mean_coi)
}
\arguments{
\item{pos}{dataframe <factor><numeric>; the genomic coordinates for chromosome and position of the sites}

\item{K}{numeric; The number of individuals (e.g. demes) to consider}

\item{m}{numeric; Probability of migration where m represents the probability of moving from \eqn{deme_{origin}} to  \eqn{deme_{new}} by \eqn{m*(1-1/K)}}

\item{rho}{numeric; expected recombination rate}

\item{mean_coi}{numeric; The lambda of a right-shifted Poisson process, \eqn{1 + Pos(lambda)} representing the average COI of an individual deme}
}
\value{
Returns a list of length G (the number of generations it took for all lineages to coalesce). Each
list item is a matrix with n x L dimensions where n is the number of parasites considered and L is loci.
Parasites are then indexed into demes (individuals) based on the COI distribution.
}
\description{
Simulate a population forwards with recombination that approximates the Structured Wright Fisher Process
and tracks haplotype identity by descent where individuals represent demes, such that within
a deme individual-level COI is considered
}
\details{
Data is simulated under the following framework:
\enumerate{
\item ... TODO>..
}
}
\examples{
\dontrun{

}

}
